@@ -1,24 +1,31 @@
 {
   "dependencies": {
-    "@emotion/core": "^10.0.27",
-    "@emotion/styled": "^10.0.27",
-    "@tippy.js/react": "^3.1.1",
-    "@types/react-dom": "^16.8.2",
-    "@types/react-helmet": "^5.0.7",
-    "gatsby": "^2.0.19",
-    "gatsby-image": "^2.0.15",
-    "gatsby-plugin-emotion": "^4.1.21",
+    "@loadable/component": "^5.14.1",
+    "@rehooks/local-storage": "^2.4.0",
+    "@types/classnames": "^2.2.11",
+    "@types/loadable__component": "^5.13.1",
+    "@types/react-dom": "^16.9.9",
+    "@types/react-helmet": "^6.1.0",
+    "autoprefixer": "^10.0.2",
+    "classnames": "^2.2.6",
+    "gatsby": "^2.26.1",
+    "gatsby-image": "^2.5.0",
     "gatsby-plugin-facebook-pixel": "^1.0.5",
-    "gatsby-plugin-google-gtag": "^1.0.8",
-    "gatsby-plugin-manifest": "^2.0.5",
-    "gatsby-plugin-react-helmet": "^3.0.0",
-    "gatsby-plugin-sharp": "^2.0.7",
-    "gatsby-plugin-typescript": "^2.0.1",
-    "gatsby-source-filesystem": "^2.0.4",
-    "gatsby-transformer-sharp": "^2.1.4",
-    "react": "^16.5.1",
-    "react-dom": "^16.5.1",
-    "react-helmet": "^5.2.0"
+    "gatsby-plugin-google-gtag": "^2.2.0",
+    "gatsby-plugin-manifest": "^2.6.0",
+    "gatsby-plugin-postcss": "^3.2.0",
+    "gatsby-plugin-react-helmet": "^3.4.0",
+    "gatsby-plugin-sharp": "^2.8.0",
+    "gatsby-plugin-typescript": "^2.6.0",
+    "gatsby-source-filesystem": "^2.5.0",
+    "gatsby-transformer-sharp": "^2.6.0",
+    "postcss": "^8.1.8",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-helmet": "^6.1.0",
+    "reactjs-popup": "^2.0.4",
+    "tailwindcss": "^2.0.1",
+    "typescript": "^4.1.2"
   },
   "name": "sadlilblackboy",
   "private": true,
