@@ -1,8 +1,8 @@
 import React from "react";
-import { Layout } from "../components/layout";
+import { Layout } from "../components/Layout";
 
 export default () => (
   <Layout>
-    <h1>NOT FOUND</h1>
+    <h1 className="text-2xl">NOT FOUND</h1>
   </Layout>
 );
